Split "comma separated values" parsing from mntent.h plumbing.
